+++ backend/api/serializers/__init__.py
@@ -1,0 +1,7 @@
+from api.serializers.BoardSerializer import BoardSerializer
+from api.serializers.ParticipantSerializer import ParticipantSerializer
+from api.serializers.PostSerializer import PostSerializer
+from api.serializers.ThreadSerializer import ThreadSerializer
+from api.serializers.UserSerializer import UserSerializer
+
+__all__ = ['BoardSerializer', 'ParticipantSerializer', 'PostSerializer', 'ThreadSerializer', 'UserSerializer']
